p2p/requirements.txt
gossip-python
redis
